plant-disease-detection-vertexai/requirements.txt
Flask
Flask-CORS
gunicorn
vertexai
google-cloud
google-oauth2-tool
requests